refactor(ui): Extract `useEnabledConnections` hook
