+++ packages/ui/src/components/sign-in/sign-in.tsx
@@ -1,7 +1,6 @@
 import { useClerk } from '@clerk/clerk-react';
 import * as Common from '@clerk/elements/common';
 import * as SignIn from '@clerk/elements/sign-in';
-import type { EnvironmentResource } from '@clerk/types';
 
 import { Connections } from '~/common/connections';
 import { EmailField } from '~/common/email-field';
@@ -15,6 +14,7 @@ import { PhoneNumberOrUsernameField } from '~/common/phone-number-or-username-fi
 import { UsernameField } from '~/common/username-field';
 import { useAttributes } from '~/hooks/use-attributes';
 import { useDisplayConfig } from '~/hooks/use-display-config';
+import { useEnabledConnections } from '~/hooks/use-enabled-connections';
 import { useLocalizations } from '~/hooks/use-localizations';
 import { Alert } from '~/primitives/alert';
 import { Button } from '~/primitives/button';
@@ -23,7 +23,6 @@ import * as Icon from '~/primitives/icon';
 import { LinkButton } from '~/primitives/link-button';
 import { SecondaryButton } from '~/primitives/secondary-button';
 import { Seperator } from '~/primitives/seperator';
-import { getEnabledSocialConnectionsFromEnvironment } from '~/utils/getEnabledSocialConnectionsFromEnvironment';
 
 export function SignInComponent() {
   return (
@@ -35,10 +34,8 @@ export function SignInComponent() {
 
 export function SignInComponentLoaded() {
   const clerk = useClerk();
-  const enabledConnections = getEnabledSocialConnectionsFromEnvironment(
-    (clerk as any)?.__unstable__environment as EnvironmentResource,
-  );
   const locationBasedCountryIso = (clerk as any)?.__internal_country;
+  const enabledConnections = useEnabledConnections();
   const { t } = useLocalizations();
   const { enabled: usernameEnabled } = useAttributes('username');
   const { enabled: phoneNumberEnabled } = useAttributes('phone_number');
