+++ packages/ui/src/hooks/use-enabled-connections.ts
@@ -1,0 +1,18 @@
+import { useClerk } from '@clerk/clerk-react';
+import { type EnvironmentResource, OAUTH_PROVIDERS, WEB3_PROVIDERS } from '@clerk/types';
+
+export function useEnabledConnections() {
+  const clerk = useClerk();
+
+  const { socialProviderStrategies, web3FirstFactors, authenticatableSocialStrategies } = (
+    (clerk as any)?.__unstable__environment as EnvironmentResource
+  ).userSettings;
+
+  const enabledConnections = new Set([
+    ...socialProviderStrategies,
+    ...web3FirstFactors,
+    ...authenticatableSocialStrategies,
+  ]);
+
+  return [...OAUTH_PROVIDERS, ...WEB3_PROVIDERS].filter(provider => enabledConnections.has(provider.strategy));
+}
